feat: chart destroy (#798)
`helm uninstall` of the otomi chart is now removing all resources by calling `otomi destroy` without the `--full` flag (which would also remove CRDs).
