+++ adr/index.md
@@ -2,7 +2,7 @@
 
 This log lists the architectural decisions for otomi-core.
 
-<!-- adrlog -- Regenerate the content by using "adr-log -i". You can install it via "npm install -g adr-log" -->
+<!-- adrlog -- Regenerate the content by using "npm run adr". -->
 
 - [ADR-2020-10-01](2020-10-01-github-workflow.md) - Our GitHub workflow
 - [ADR-2020-10-02](2020-10-02-docker-compose.md) - docker compose
@@ -10,8 +10,10 @@ This log lists the architectural decisions for otomi-core.
 - [ADR-2020-11-06](2020-11-06-konstraint-policy-library.md) - Konstraint library to support Common Rego Language definitions
 - [ADR-2021-10-18](2021-10-18-defaults-and-derived.md) - Derived values
 - [ADR-2021-10-28](2021-10-28-internal-values.md) - Internal values
+- [ADR-2022-03-24](2022-03-24-custom-ca.md) - urice:
 - [ADR-2022-04-22](2022-04-22-values-migration.md) - Values migration
 - [ADR-2022-04-23](2022-04-23-pre-upgrade.md) - A new otomi pre-upgrade command
+- [ADR-2022-05-17](2022-05-17-destroy-upon-uninstall.md) - Extra flags to accomodate destroy upon uninstall
 
 <!-- adrlogstop -->
