+++ chart/otomi/localtest.sh
@@ -4,4 +4,5 @@ export VALUES_INPUT=${VALUES_INPUT:-'/tmp/otomi/values.yaml'}
 export DOCKER_EXTRA_ARGS="-v $(dirname $VALUES_INPUT):$(dirname $VALUES_INPUT)"
 export VERBOSITY=${VERBOSITY:-'1'}
 
-binzx/otomi bootstrap -t
+binzx/otomi bootstrap -tv
+binzx/otomi apply -v
