+++ chart/otomi/templates/post-job.yaml
@@ -1,3 +1,4 @@
+{{- $kms := .Values.kms | default dict }}
 {{- $version := .Values.otomi.version | default .Chart.AppVersion }}
 apiVersion: batch/v1
 kind: Job
@@ -8,32 +9,55 @@ metadata:
     helm.sh/hook-delete-policy: hook-succeeded
   labels: {{- include "otomi.labels" . | nindent 4 }}
 spec:
+  backoffLimit: 2
   ttlSecondsAfterFinished: 100
   template:
     metadata:
-      labels:
-        app: {{ template "otomi.name" . }}
-        release: {{ .Release.Name }}
+      {{- with .Values.podAnnotations }}
+      annotations: {{- toYaml . | nindent 8 }}
+      {{- end }}
+      labels: {{- include "otomi.selectorLabels" . | nindent 8 }}
     spec:
       restartPolicy: Never
       serviceAccountName: {{ include "otomi.fullname" . }}
+      securityContext:
+        runAsUser: 999
+        runAsGroup: 999
       containers:
-        - name: pre-delete
-          image: bitnami/kubectl:{{ .Values.cluster.k8sVersion }}
-          command:
-            - sh
-            - -c
-          args:
-            - |
-              kubectl delete cm otomi-status
-              kubectl delete secret otomi-generated-passwords
-              echo "Otomi has been uninstalled."
+        - name: destroy
+          image: otomi/core:{{ $version }}
+          imagePullPolicy: {{ ternary "IfNotPresent" "Always" (regexMatch "^v\\d" $version) }} 
           resources:
             limits:
-              memory: 100Mi
-              cpu: 100m
+              memory: 2Gi
+              cpu: '2'
             requests:
-              memory: 50Mi
-              cpu: 50m
-          securityContext:
-            runAsUser: 999
+              memory: 1Gi
+              cpu: '1'
+          command: [bash, -c]
+          args:
+            - |
+              binzx/otomi bootstrap --destroy && binzx/otomi destroy && echo "Otomi has been uninstalled."
+          env:
+            - name: VERBOSITY
+              value: '1'
+            - name: ENV_DIR
+              value: /home/app/stack/env
+            - name: VALUES_INPUT
+              value: /secret/values.yaml
+          {{- if hasKey $kms "sops" }}
+          envFrom:
+            - secretRef:
+                name: {{ include "otomi.fullname" . }}-sops-secrets
+          {{- end }}
+          volumeMounts:
+            - name: otomi-values
+              mountPath: /home/app/stack/env
+            - name: values-secret
+              mountPath: /secret
+      volumes:
+        - name: values-secret
+          secret:
+            secretName: {{ .Release.Name }}-values
+        - name: otomi-values
+          emptyDir: {}
