+++ src/cmd/apply.ts
@@ -1,14 +1,14 @@
 import { mkdirSync, rmdirSync, writeFileSync } from 'fs'
-import { isIPv6 } from 'net'
 import { Argv, CommandModule } from 'yargs'
 import { $, nothrow } from 'zx'
+import { prepareDomainSuffix } from '../common/bootstrap'
 import { cleanupHandler, prepareEnvironment } from '../common/cli'
 import { logLevelString, terminal } from '../common/debug'
 import { env, isChart } from '../common/envalid'
-import { hf, hfValues } from '../common/hf'
-import { getDeploymentState, getOtomiLoadBalancerIP, setDeploymentState } from '../common/k8s'
+import { hf } from '../common/hf'
+import { getDeploymentState, setDeploymentState } from '../common/k8s'
 import { getFilename } from '../common/utils'
-import { getCurrentVersion, getImageTag, writeValues } from '../common/values'
+import { getCurrentVersion, getImageTag } from '../common/values'
 import { getParsedArgs, HelmArguments, helmOptions, setParsedArgs } from '../common/yargs'
 import { ProcessOutputTrimmed } from '../common/zx-enhance'
 import { commit } from './commit'
@@ -26,32 +26,6 @@ const setup = (): void => {
   const argv: HelmArguments = getParsedArgs()
   cleanupHandler(() => cleanup(argv))
   mkdirSync(dir, { recursive: true })
-}
-
-const setDomainSuffix = async (values: Record<string, any>): Promise<void> => {
-  const d = terminal(`cmd:${cmdName}:setDomainSuffix`)
-  d.debug("Create a fallback cluster.domainSuffix when it doesn't exist")
-  const ingressIP = values.apps['ingress-nginx']?.loadBalancerIP ?? (await getOtomiLoadBalancerIP())
-  // When ingressIP is V6, we need to use sslip.io as they resolve it, otherwise use nip.io as it uses PowerDNS
-  const newSuffix = isIPv6(ingressIP) ? `${ingressIP.replaceAll(':', '-')}.sslip.io` : `${ingressIP}.nip.io`
-
-  await writeValues({
-    cluster: {
-      domainSuffix: newSuffix,
-    },
-  })
-}
-
-const prepareValues = async (): Promise<void> => {
-  const d = terminal(`cmd:${cmdName}:prepareValues`)
-
-  const values = await hfValues()
-  if (!values) throw new Error('No values???')
-  d.info('Checking if domainSuffix needs a fallback domain')
-  if (values && !values.cluster.domainSuffix) {
-    d.info('cluster.domainSuffix was not found, creating $loadbalancerIp.nip.io as fallback')
-    await setDomainSuffix(values)
-  }
 }
 
 const applyAll = async () => {
@@ -91,7 +65,7 @@ const applyAll = async () => {
     },
     { streams: { stdout: d.stream.log, stderr: d.stream.error } },
   )
-  await prepareValues()
+  await prepareDomainSuffix()
   d.info('Deploying charts containing label stage!=prep')
   await hf(
     {
