+++ src/cmd/commit.ts
@@ -1,13 +1,14 @@
-import { copyFileSync, existsSync } from 'fs-extra'
 import { Argv } from 'yargs'
-import { $, cd, nothrow } from 'zx'
+import { $, cd } from 'zx'
+import { bootstrapGit } from '../common/bootstrap'
 import { prepareEnvironment } from '../common/cli'
 import { encrypt } from '../common/crypt'
 import { terminal } from '../common/debug'
-import { env, isCli } from '../common/envalid'
+import { env } from '../common/envalid'
 import { hfValues } from '../common/hf'
 import { setDeploymentState, waitTillAvailable } from '../common/k8s'
-import { getFilename, rootDir } from '../common/utils'
+import { getFilename } from '../common/utils'
+import { getRepo } from '../common/values'
 import { getParsedArgs, HelmArguments, setParsedArgs } from '../common/yargs'
 import { Arguments as DroneArgs, genDrone } from './gen-drone'
 import { validateValues } from './validate-values'
@@ -19,69 +20,8 @@ interface Arguments extends HelmArguments, DroneArgs {
   message?: string
 }
 
-export const bootstrapGit = async (inValues?: Record<string, any>): Promise<void> => {
-  let values = inValues
-  if (!values) {
-    if (isCli) values = (await hfValues({ filesOnly: true })) as Record<string, any>
-    else return
-  }
-  if (!values?.cluster?.domainSuffix) return // too early, commit will handle it
-  const d = terminal(`cmd:${cmdName}:bootstrapGit`)
-  const { remote, branch, email, username, password } = getRepo(values)
-  cd(env.ENV_DIR)
-  process.env.GIT_SSL_NO_VERIFY = '1' // we don't care about this as repo endpoint is either ours or user input
-  let hasCommits = false
-  try {
-    d.debug('Checking out remote into /tmp/xx to test if repo exists and use if needed')
-    // check remote exists by cloning with a 10 second timeout (if remote is unreachable it takes 30 secs to timeout)
-    await $`set +e && rm -rf /tmp/xx >/dev/null && set -e && timeout 10 git clone ${remote} /tmp/xx`
-    // it didn't throw, so we know we have an existing remote
-    // do we have commits remotely?
-    try {
-      await $`cd /tmp/xx && git fetch && git checkout ${branch} && git log && cd -`
-      d.info('We have commits, so we must clone first.')
-      hasCommits = true
-    } catch (e) {
-      d.error(e)
-      cd(env.ENV_DIR) // cd back to be nice
-      if (!`${e}`.includes('would be overwritten by checkout')) {
-        d.info('No commits found, We should be ok to push.')
-        return
-      }
-    }
-    // we know we have commits, so we use the clone and rsync local files if these don't exist yet
-    try {
-      await $`rsync -a --ignore-existing . /tmp/xx/`
-      await $`rm -rf .[!.]* *  && rsync -av --no-o --no-g --no-p /tmp/xx/ ./`
-    } catch (e) {
-      d.error(`An error occured when trying to sync with the clone. This is a fatal error: ${e}`)
-    }
-  } catch (e) {
-    d.log(e)
-    $.verbose = false
-    d.info('Remote does not exist yet. Expecting first commit to come later.')
-  }
-
-  if (!existsSync(`${env.ENV_DIR}/.git`)) {
-    d.info('Initializing values git repo.')
-    await $`git init ${env.ENV_DIR}`
-  }
-  if (isCli) copyFileSync(`${rootDir}/bin/hooks/pre-commit`, `${env.ENV_DIR}/.git/hooks/pre-commit`)
-  else await nothrow($`git config --global --add safe.directory ${env.ENV_DIR}`)
-  await nothrow($`git config --local user.name ${username}`)
-  await nothrow($`git config --local user.password ${password}`)
-  await nothrow($`git config --local user.email ${email}`)
-  if (!hasCommits) {
-    await nothrow($`git checkout -b ${branch}`)
-    await nothrow($`git remote add origin ${remote}`)
-  }
-  if (existsSync(`${env.ENV_DIR}/.sops.yaml`)) await nothrow($`git config --local diff.sopsdiffer.textconv "sops -d"`)
-  d.log(`Done bootstrapping git`)
-}
-
-const gitPush = async (): Promise<boolean> => {
+const gitPush = async (values: Record<string, any>): Promise<boolean> => {
   const d = terminal(`cmd:${cmdName}:gitPush`)
-  const values = (await hfValues()) as Record<string, any>
   let branch = 'main'
   if (values.apps?.gitea?.enabled === false) {
     branch = values.apps!['otomi-api']!.git!.branch ?? branch
@@ -98,7 +38,7 @@ const gitPush = async (): Promise<boolean> => {
   }
 }
 
-const commitAndPush = async (): Promise<void> => {
+const commitAndPush = async (values: Record<string, any>): Promise<void> => {
   const d = terminal(`cmd:${cmdName}:commitAndPush`)
   d.info('Committing values')
   const argv = getParsedArgs()
@@ -112,44 +52,18 @@ const commitAndPush = async (): Promise<void> => {
     d.log('Something went wrong trying to commit. Did you make any changes?')
   }
   try {
+    const giteaEnabled = values?.apps?.gitea?.enabled
+    const byor = !!values?.apps?.['otomi-api']?.git
+    if (giteaEnabled && !byor && values._derived?.untrustedCA) {
+      process.env.GIT_SSL_NO_VERIFY = '1'
+    }
     await $`git remote show origin`
-    await gitPush()
+    await gitPush(values)
     d.log('Successfully pushed the updated values')
   } catch (error) {
     d.error(error.stderr)
     throw new Error('Pushing the values failed, please read the above error message and manually try again')
   }
-}
-
-export const getRepo = (values): Record<string, any> => {
-  const giteaEnabled = values?.apps?.gitea?.enabled ?? true
-  const clusterDomain = values?.cluster?.domainSuffix
-  const byor = !!values?.apps?.['otomi-api']?.git
-  if (!giteaEnabled && !byor) {
-    throw new Error('Gitea is disabled but no apps.otomi-api.git config was given.')
-  }
-  let username = 'Otomi Admin'
-  let email: string
-  let password: string
-  let branch = 'main'
-  let remote
-  if (!giteaEnabled) {
-    const otomiApiGit = values?.apps?.['otomi-api']?.git
-    username = otomiApiGit?.user
-    password = otomiApiGit?.password
-    remote = otomiApiGit?.repoUrl
-    email = otomiApiGit?.email
-    branch = otomiApiGit?.branch ?? branch
-  } else {
-    username = 'otomi-admin'
-    password = values?.apps?.gitea?.adminPassword ?? values?.otomi?.adminPassword
-    email = `otomi-admin@${clusterDomain}`
-    const giteaUrl = `gitea.${clusterDomain}`
-    const giteaOrg = 'otomi'
-    const giteaRepo = 'values'
-    remote = `https://${username}:${encodeURIComponent(password)}@${giteaUrl}/${giteaOrg}/${giteaRepo}.git`
-  }
-  return { remote, branch, email, username, password }
 }
 
 export const commit = async (firstTime = false): Promise<void> => {
@@ -158,6 +72,7 @@ export const commit = async (firstTime = false): Promise<void> => {
   d.info('Preparing values')
   const values = (await hfValues()) as Record<string, any>
   const { remote } = getRepo(values)
+  // we call this here again, as we might not have completed (happens upon first install):
   await bootstrapGit(values)
   if (values?.apps!.gitea!.enabled) {
     const { adminPassword } = values.apps!.gitea
@@ -170,7 +85,7 @@ export const commit = async (firstTime = false): Promise<void> => {
   }
   await genDrone()
   await encrypt()
-  await commitAndPush()
+  await commitAndPush(values)
 
   if (firstTime) {
     await setDeploymentState({ status: 'deployed' })
