+++ src/cmd/destroy.ts
@@ -5,11 +5,15 @@ import { cleanupHandler, prepareEnvironment } from '../common/cli'
 import { logLevelString, terminal } from '../common/debug'
 import { hf } from '../common/hf'
 import { getFilename } from '../common/utils'
-import { getParsedArgs, HelmArguments, helmOptions, setParsedArgs } from '../common/yargs'
+import { BasicArguments, getParsedArgs, HelmArguments, helmOptions, setParsedArgs } from '../common/yargs'
 import { ProcessOutputTrimmed, stream } from '../common/zx-enhance'
 
 const cmdName = getFilename(__filename)
 const templateFile = '/tmp/otomi/destroy-template.yaml'
+
+export interface Arguments extends BasicArguments {
+  full?: boolean
+}
 
 const cleanup = (argv: HelmArguments): void => {
   if (argv.skipCleanup) return
@@ -21,6 +25,7 @@ const setup = (argv: HelmArguments): void => {
 }
 
 const destroyAll = async () => {
+  const argv = getParsedArgs()
   const d = terminal(`cmd:${cmdName}:destroyAll`)
   d.log('Uninstalling otomi...')
   const debugStream = { stdout: d.stream.debug, stderr: d.stream.error }
@@ -46,8 +51,8 @@ const destroyAll = async () => {
   writeFileSync(templateFile, templateOutput)
   await stream(nothrow($`kubectl delete -f ${templateFile}`), debugStream)
   d.info('Uninstalled all manifests.')
+  if (!argv.full) return
   d.info('Uninstalling CRDs...')
-
   const ourCRDS = [
     'argoproj.io',
     'appgw.ingress.k8s.io',
@@ -105,7 +110,15 @@ const destroy = async (): Promise<void> => {
 export const module = {
   command: cmdName,
   describe: 'Destroy all, or supplied, k8s resources',
-  builder: (parser: Argv): Argv => helmOptions(parser),
+  builder: (parser: Argv): Argv =>
+    helmOptions(
+      parser.options({
+        full: {
+          boolean: true,
+          default: false,
+        },
+      }),
+    ),
 
   handler: async (argv: HelmArguments): Promise<void> => {
     setParsedArgs(argv)
