+++ src/cmd/values.ts
@@ -8,15 +8,16 @@ import { BasicArguments, getParsedArgs, setParsedArgs } from '../common/yargs'
 
 const cmdName = getFilename(__filename)
 
-export interface Arguments extends BasicArguments {
+interface Arguments extends BasicArguments {
   filesOnly?: boolean
+  excludeSecrets?: boolean
 }
 
 const values = async (): Promise<void> => {
   const d = terminal(`cmd:${cmdName}:values`)
   d.info('Get values')
   const argv: Arguments = getParsedArgs()
-  const hfVal = await hfValues({ filesOnly: argv.filesOnly })
+  const hfVal = await hfValues({ filesOnly: argv.filesOnly, excludeSecrets: argv.excludeSecrets })
 
   d.info('Print values')
   console.log(dump(hfVal))
@@ -24,10 +25,15 @@ const values = async (): Promise<void> => {
 
 export const module = {
   command: cmdName,
-  describe: 'Show helmfile values for target cluster (--files-only: only values stored on disk)',
+  describe:
+    'Show helmfile values for target cluster (--files-only: only values stored on disk, --exclude-secrets: omit secrets)',
   builder: (parser: Argv): Argv =>
     parser.options({
       filesOnly: {
+        boolean: true,
+        default: false,
+      },
+      excludeSecrets: {
         boolean: true,
         default: false,
       },
