+++ src/common/constants.ts
@@ -1,3 +1,2 @@
 export const DEPLOYMENT_PASSWORDS_SECRET = 'otomi-generated-passwords'
-export const DEPLOYMENT_INFO_SECRET = 'otomi-deployment-info'
 export const DEPLOYMENT_STATUS_CONFIGMAP = 'otomi-status'
